refactor: extract FixEventList, sticky PaneHeader, minor polish
- Extract fix-event list renderer into shared FixEventList component
  (used by both CommentsLayer thread popover and CommentsPane)
- Make PaneHeader sticky so section titles stay visible while scrolling
- Cache getFixProgress snapshot between change events
- Skip redundant markDirty/requestLayout in addAnnotationReply when
  updateAnnotationStatus already handled them
- Hide MCP Server setup section in DocumentPane (CLI-only flow)
- Polish Fix settings copy ("Bypass permissions", text-pretty wrap)
