+++ src/renderer/above-view/CommentsLayer.tsx
@@ -1,8 +1,9 @@
-import { Fragment, useEffect, useRef } from 'react'
-import type { Annotation, FixProgressEntry, FixProgressEvent, LayoutUpdateData, WorkspaceBounds } from '../../shared/types'
+import { useRef } from 'react'
+import type { Annotation, FixProgressEntry, LayoutUpdateData, WorkspaceBounds } from '../../shared/types'
 import { canvasRectToScreenRect, type PendingAnnotation } from './annotationMath'
 import { CircleCheckIcon, MoreVerticalIcon, TrashIcon } from '../shared/PanelIcons'
 import { CommentBubble, CommentInput } from '../shared/CommentPrimitives'
+import { FixEventList } from '../shared/FixEventList'
 
 export function PendingCommentComposer({
   clearDraft,
@@ -267,14 +268,7 @@ export function AnnotationThreadPopover({
 }
 
 function ThreadFixProgress({ progress }: { progress: FixProgressEntry }) {
-  const scrollRef = useRef<HTMLDivElement | null>(null)
   const eventCount = progress.events.length
-
-  useEffect(() => {
-    const node = scrollRef.current
-    if (!node) return
-    node.scrollTop = node.scrollHeight
-  }, [eventCount])
 
   const statusLabel =
     progress.status === 'running'
@@ -290,16 +284,6 @@ function ThreadFixProgress({ progress }: { progress: FixProgressEntry }) {
         ? 'text-red-600 dark:text-red-400'
         : 'text-emerald-600 dark:text-emerald-400'
 
-  const kindColor: Record<FixProgressEvent['kind'], string> = {
-    system: 'text-zinc-500',
-    text: 'text-zinc-800 dark:text-zinc-200',
-    tool_use: 'text-blue-700 dark:text-blue-300',
-    tool_result: 'text-emerald-700 dark:text-emerald-300',
-    result: 'text-zinc-900 dark:text-zinc-100',
-    stderr: 'text-amber-700 dark:text-amber-300',
-    error: 'text-red-700 dark:text-red-300',
-  }
-
   return (
     <div className="border-t border-zinc-200 dark:border-zinc-700">
       <div className="flex items-center justify-between px-2.5 py-1.5 text-[11px]">
@@ -309,21 +293,7 @@ function ThreadFixProgress({ progress }: { progress: FixProgressEntry }) {
         </span>
       </div>
       {eventCount > 0 ? (
-        <div
-          ref={scrollRef}
-          className="grid max-h-[160px] auto-rows-min grid-cols-[auto_minmax(0,1fr)] gap-x-2 gap-y-0 overflow-auto px-2.5 pb-2 font-mono text-[11px] leading-relaxed"
-        >
-          {progress.events.map((event, i) => (
-            <Fragment key={`${event.timestamp}-${i}`}>
-              <span className="text-zinc-400 dark:text-zinc-600">
-                {event.kind.replace('_', ' ')}
-              </span>
-              <span className={`break-words ${kindColor[event.kind]}`}>
-                {event.text}
-              </span>
-            </Fragment>
-          ))}
-        </div>
+        <FixEventList events={progress.events} className="max-h-[160px] px-2.5 pb-2" />
       ) : (
         <div className="px-2.5 pb-2 text-[11px] text-zinc-400 dark:text-zinc-500">
           Waiting for output…
